pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = ""
authors = ["pe4eniks <evgeniipishchik@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "^3.5.0"
torch = "2.1.1"
dvc = {extras = ["gdrive"], version = "^3.30.3"}
numpy = "^1.26.2"
tqdm = "^4.66.1"
scikit-learn = "^1.3.2"
torchvision = "^0.16.1"
torchmetrics = "^1.2.1"
hydra-core = "^1.3.2"
onnx = "^1.15.0"
onnxscript = "^0.1.0.dev20231213"
onnxruntime = "^1.16.3"
pytorch-lightning = "^2.1.3"
mlflow = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
